Refactor project - Move components inside of respective pages - Create folders for each page - Remove unncessary comments - Remove files not in use
