+++ frontend/src/App.jsx
@@ -1,15 +1,19 @@
 import React, { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import './App.css';
+
+// Pages
+import Index from './Pages/Index/Index';
+import AboutUs from './Pages/AboutUs/AboutUs'
+import Splash from './Pages/Splash/Splash';
+import CreatePet from './Pages/CreatePet/CreatePet';
+import EditPet from './Pages/EditPet/EditPet';
+import UserProfile from './Pages/UserProfile/UserProfile';
+import ShowPet from './Pages/ShowPet/ShowPet';
+
+// Components
 import NavBar from './Components/NavBar/NavBar';
 import Footer from './Components/Footer/Footer';
-import Index from './Pages/Index/Index';
-import AboutUs from './Pages/AboutUs';
-import Splash from './Pages/Splash';
-import CreatePet from './Pages/CreatePet';
-import EditPet from './Pages/EditPet';
-import UserProfile from './Pages/UserProfile';
-import ShowPet from './Pages/ShowPet';
-import './App.css';
 
 function App() {
   return (
@@ -19,10 +23,7 @@ function App() {
         <div className="content">
           <Routes>
             <Route path="/" element={<Splash />} />
-            <Route
-              path="/index"
-              element={<Index />}
-            />
+            <Route path="/index" element={<Index />} />
             <Route path="/pets/:id" element={<ShowPet />} />
             <Route path="/create-pet" element={<CreatePet />} />
             <Route path="/edit-pet/:id" element={<EditPet />} />
