--- frontend/src/Pages/CreatePet.scss
@@ -1,46 +1,0 @@
-.create-pet {
-    max-width: 600px;
-    margin: 0 auto;
-    padding: 20px;
-    background-color: #f9f9f9;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  }
-  
-  .create-pet h2 {
-    text-align: center;
-    margin-bottom: 20px;
-  }
-  
-  .create-pet form {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .create-pet label {
-    margin-bottom: 10px;
-    font-weight: bold;
-  }
-  
-  .create-pet input,
-  .create-pet textarea {
-    padding: 10px;
-    margin-top: 5px;
-    margin-bottom: 15px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-  }
-  
-  .create-pet button {
-    padding: 10px;
-    border: none;
-    border-radius: 4px;
-    background-color: #007bff;
-    color: white;
-    cursor: pointer;
-  }
-  
-  .create-pet button:hover {
-    background-color: #0056b3;
-  }
-
