+++ frontend/src/Pages/CreatePet/CreatePet.jsx
@@ -1,0 +1,9 @@
+import CreateForm from '../../Components/CreateForm/CreateForm';
+
+export default function CreatePet() {
+  return (
+    <div className="CreatePet">
+      <CreateForm/>
+    </div>
+  )
+}
