--- frontend/src/Pages/EditPet.scss
@@ -1,46 +1,0 @@
-.edit-pet {
-    max-width: 600px;
-    margin: 0 auto;
-    padding: 20px;
-    background-color: #f9f9f9;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  }
-  
-  .edit-pet h2 {
-    text-align: center;
-    margin-bottom: 20px;
-  }
-  
-  .edit-pet form {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .edit-pet label {
-    margin-bottom: 10px;
-    font-weight: bold;
-  }
-  
-  .edit-pet input,
-  .edit-pet textarea {
-    padding: 10px;
-    margin-top: 5px;
-    margin-bottom: 15px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-  }
-  
-  .edit-pet button {
-    padding: 10px;
-    border: none;
-    border-radius: 4px;
-    background-color: #007bff;
-    color: white;
-    cursor: pointer;
-  }
-  
-  .edit-pet button:hover {
-    background-color: #0056b3;
-  }
-
