+++ frontend/src/Pages/EditPet/EditPet.jsx
@@ -1,0 +1,9 @@
+import EditForm from '../../Components/EditForm/EditForm';
+
+export default function EditPet() {
+  return (
+    <div className="EditForm">
+      <EditForm />
+    </div>
+  );
+}
