+++ frontend/src/Pages/Splash/Splash.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import NavBar from '../Components/NavBar/NavBar';
-import Hero from '../Components/Hero/Hero';
-import Footer from '../Components/Footer/Footer';
+import NavBar from '../../Components/NavBar/NavBar';
+import Hero from '../../Components/Hero/Hero';
+import Footer from '../../Components/Footer/Footer';
 
 const Splash = () => {
   return (
@@ -10,7 +10,6 @@ const Splash = () => {
       <div className="splash-content">
         {/* Add any additional content or components here */}
       </div>
-      
     </div>
   );
 };
