+++ frontend/src/Pages/Splash/Splash.scss
