--- frontend/variables.scss
@@ -1,6 +1,0 @@
-// Colors
-$textColor: #000;
-$backgroundColor: #000;
-$primaryColor: #000;
-$secondaryColor: #000;
-$accentColor: #000;
